\name{tswv}
\docType{data}
\alias{tswv}
\title{Tomato Spotted Wilt Virus data}
\description{
  Data extracted from plots presented in Hughes et al. (1997)
}
\usage{data(tswv)}
\format{
  A data frame with following variables
   \describe{
    \item{x}{ X coordinate}
    \item{y}{ Y coordinate}
    \item{inftime}{Infection times}
    \item{removaltime}{Times at which individuals are removed}
  }
}
\references{
Hughes, G., McRoberts,N., Madden, L.V., Nelson, S. C. (1997). Validating mathematical models of plant disease progress in space and time. \emph{IMA Journal of Mathematics Applied in Medicine and Biology}, 14, 85-112.
}